requirements.txt
Django==1.4
PIL==1.1.7
South==0.7.3
django-imagekit==0.3.6
wsgiref==0.1.2
django-social-auth
django_compressor
django-registration==0.8
django-tastypie
